+++ CHANGES.md
@@ -4,6 +4,7 @@ History
 ## Version 0.0.15 (next version)
 
 * Support `raf` and `mutation` polling for `waitFor*` methods
+* Add XPath support with `Page.xpath`, `Frame.xpath`, and `ElementHandle.xpath`
 * Change `Frame.executionContest` from property to coroutine
 * Implement `Frame.content` and `Frame.setContent` methods
 * `page.tracing.start` accepts custom tracing categories option